requirements.txt
asn1crypto==0.24.0
astroid==2.0.4
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
cryptography==2.3.1
defusedxml==0.5.0
Django==2.1
django-allauth==0.36.0
django-filter==2.0.0
django-rest-auth==0.9.3
django-rest-swagger==2.1.2
django-storages==1.6.6
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
drf-yasg==1.10.0
dropbox==9.0.0
flex==6.13.2
future==0.16.0
gunicorn==19.9.0
idna==2.7
inflection==0.3.1
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
jsonpointer==1.14
jsonschema==2.6.0
lazy-object-proxy==1.3.1
Markdown==2.6.11
MarkupSafe==1.0
mccabe==0.6.1
numpy==1.14.5
oauthlib==2.1.0
openapi-codec==1.3.2
pandas==0.23.4
Pillow==5.2.0
pycparser==2.18
PyJWT==1.6.4
pylint==2.1.1
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.5
PyYAML==3.13
requests==2.19.1
requests-oauthlib==1.0.0
rfc3987==1.3.8
ruamel.yaml==0.15.60
simplejson==3.16.0
six==1.11.0
strict-rfc3339==0.7
swagger-spec-validator==2.3.1
typed-ast==1.1.0
uritemplate==3.0.0
urllib3==1.23
validate-email==1.3
whitenoise==4.0
wrapt==1.10.11
